requirements.txt
streamlit
google-generativeai
pytesseract
pdf2image
spacy
pandas
Pillow
